[FIX] web_editor, website, *: improve post install theme and start tour
Before this commit, once you install website_* module, the ir.actions.todo from
the installed module was overidded by the 'choose a theme' action.

Now, when we consume this ir_act_todo, we re-open the open_menu that contains
most of the time the expected next action from new module installed.

Since we can also install module from the '+ new' button, in frontend, we
considere that it first step of tour is to click on it, it is already done and
we skip it.

Install website_event
Before:
   -> not redirected to /event (the ir.action.todo)
   -> no tour, if you don't click manually on '+ new'

After:
    -> you are redirected to /event
    -> you can launch tour from '+ new' if you are alrady on the new panel,
       start direclty to the second step 'create event'